Code rearranged
