+++ app/src/record_module.h
@@ -1,1 +1,2 @@
+#pragma once
